update expandable block
