+++ apps/website/src/content/docs/expandableblock.mdx
@@ -108,6 +108,18 @@ Use a [hierarchy tree](tree) when:
 - There are only list items in the pattern
 - The hierarchy is not necessarily linear and may branch out
 
-## Props
+## ExpandableBlock
 
 <PropsTable path='@itwin/itwinui-react/esm/core/ExpandableBlock/ExpandableBlock.d.ts' />
+
+## ExpandableBlock.Wrapper
+
+<PropsTable path='@itwin/itwinui-react/esm/core/ExpandableBlock/ExpandableBlockWrapper.d.ts' />
+
+## ExpandableBlock.Content
+
+<PropsTable path='@itwin/itwinui-react/esm/core/ExpandableBlock/ExpandableBlockContent.d.ts' />
+
+## ExpandableBlock.Trigger
+
+<PropsTable path='@itwin/itwinui-react/esm/core/ExpandableBlock/ExpandableBlockTrigger.d.ts' />
